+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 # Unreleased
 
+- Make `deprecated.note` optional for `{reason: renamed}` - inferred from `renamed_to`. ([#1622](https://github.com/open-telemetry/weaver/pull/1622) by @lmolkova)
 - Add v2 entity refinements to the resolved and materialized schema, allow to refine attributes without changing entity identity. ([#1588](https://github.com/open-telemetry/weaver/pull/1588) by @lmolkova)
 - Live-check: (Fixes: [#1613](https://github.com/open-telemetry/weaver/issues/1613)) add `sample_names` to `[[live-check.finding_filters]]` to scope a filter to matching sample names, with glob wildcard support (also added to `exclude_samples`). ([#1619](https://github.com/open-telemetry/weaver/pull/1619) by @jerbly)
 - Add a tree view to the `serve` UI's search page, grouping results by namespace with expand/collapse controls, and a "Hide deprecated" toggle (on by default) for both the list and tree views. ([#1595](https://github.com/open-telemetry/weaver/pull/1595) by @jerbly)